+++ src/application/mod.rs
@@ -5,6 +5,7 @@
 pub mod account_order;
 pub mod accounts;
 pub mod allowed;
+pub mod answered_meetings;
 pub mod answering;
 pub mod asking_for_a_whole_folder;
 pub mod asking_when_free;